[Gold IV] Title: 준규와 사과, Time: 116 ms, Memory: 14896 KB -BaekjoonHub

+++ 백준/Gold/5913. 준규와 사과/준규와 사과.java
@@ -21,7 +21,7 @@ public class Main {
         visited[0][0] = true;
         arr[4][4] = 0;
         visited[4][4] = true;
-        long rst = backtracking(0, 0, 4, 4, 25 - k - 2);
+        int rst = backtracking(0, 0, 4, 4, 25 - k - 2);
         System.out.println(rst);
     }
